feat: Add MoE plugin conversion and optimization
- Implement MixtureOfExpertsPlugin and its inputs
- Implement MoEConfig for MoE plugin fields
- Enhance MoESubgraph to support plugin replacement and unused node removal
- Introduce plugin converter for MoE layers in conversion pipeline

+++ src/ditto/fx/optimize.py
@@ -207,7 +207,7 @@ def get_trtllm_conversion_transform(
         AddTRTLLMInputs(argument_hint=argument_hint),
         *get_trtllm_output_adaptation_passes(model_config.gather_context_logits),
         StashLoraSubgraphs,
-        ReplaceMoEByMoEPlugin,  # TODO: consider the order
+        ReplaceMoEByMoEPlugin(dtype=dtype),
         FuseQKVProjections,
         FuseGatedMLPProjections,
         WrapRoPESubgraphs,
